爬虫重构阶段二 (#20)
打算做的：

* （OK）把 spider 里面的缓存部分搬到后端（登录状态的缓存还保留）
* （OK）~~把 spider 的 handler 砍了~~~~（打算 handler 还保留，用于解析数据）~~把 spider 的 handler 砍了，后端的 infra::spider 砍了
* （OK）把后端的数据解析部分搬到 spider
* （OK）顺便把看不下去的代码重构了
* （OK）把后端的部分测试也砍掉或者是搬到 spider
* （OK）补点注释

Reviewed-on: http://code.qnxg.cn/qnxg/back_2026/pulls/20

+++ back_2024/src/routers/exam.rs
@@ -90,8 +90,8 @@ async fn get_exam_arrange(req: &mut Request) -> RouterResult {
     let stu_id = utils::jwt::auth(req)?;
     let (current_xn, current_xq) =
         service::semester::get_now_xnxq().await?;
-    let xn = xn.unwrap_or(current_xn);
-    let xq = xq.unwrap_or(current_xq);
+    let xn = xn.unwrap_or(current_xn) as u16;
+    let xq = xq.unwrap_or(current_xq) as u8;
     let res =
         service::exam::get_exam_arrange(&stu_id, xn, xq).await?;
     Ok(res.into())
